requirements.txt
allure-behave~=2.8.31
Appium-Python-Client~=1.0.2
argparse
behave~=1.2.6
configcat-client>=5.0.0
configparser~=5.0.0
contextvars
Django
pactman>=2.28.0
pdfminer
pip
PyPDF2
python-dateutil~=2.8.1
pytz
requests
selenium~=3.141.0
setuptools
urllib3
